Some code re-organization to split large files and types into their own files
