+++ Sources/SwiftModel/Internal/TestAccess.swift
@@ -160,10 +160,10 @@ final class TestAccess<Root: Model>: ModelAccess, TaskLifecycleDelegate, @unchec
 
     // Activation task counter (Phase 3): tracks tasks created inside onActivate() that
     // have not yet begun executing their body. Reaches 0 when all such tasks have entered.
-    private var _activationTasksInFlight: Int = 0
+    var _activationTasksInFlight: Int = 0
 
     // General task counter (Phase 5): tracks all currently-running tasks in the hierarchy.
-    private var _activeTaskCount: Int = 0
+    var _activeTaskCount: Int = 0
 
 
     // Captures a single state transition: how to apply it to a Root snapshot, and how to
@@ -673,883 +673,36 @@ final class TestAccess<Root: Model>: ModelAccess, TaskLifecycleDelegate, @unchec
         }
     }
 
-    func expect(settleResetting: _ExhaustivityBits? = nil, at fileAndLine: FileAndLine, predicates: [AssertBuilder.Predicate], enableExhaustionTest: Bool = true) async {
-        await expect(timeoutNanoseconds: 1_000_000_000, settleResetting: settleResetting, at: fileAndLine, predicates: predicates, enableExhaustionTest: enableExhaustionTest)
-    }
-
-    func expect(timeoutNanoseconds timeout: UInt64, settleResetting: _ExhaustivityBits? = nil, at fileAndLine: FileAndLine, predicates: [AssertBuilder.Predicate], enableExhaustionTest: Bool = true) async {
-        let cal = await calibrate(timeout: timeout)
-        let scaledTimeout = cal.scaledTimeout
-        let yieldRoundNs = cal.yieldRoundNs
-        let hardCap = cal.hardCap
-        let context = TesterAssertContext(events: { self.lock { self.events } }, fileAndLine: fileAndLine)
-
-        await TesterAssertContextBase.$assertContext.withValue(context) {
-
-            let start = cal.start
-            var lastProgressTime = start  // reset whenever a state modification arrives
-            var retryCount = 0
-            var failures: [TesterAssertContext.Failure] = []
-            var passedAccesses: [TesterAssertContext.Access] = []
-            while true {
-                // Defensive: clear any stale transition override that might linger on this
-                // pthread thread from a previous iteration. The willAccess callback normally
-                // clears it, but skipped-callback paths (e.g. destructed model branch in
-                // Context._read) can leave it set. Clearing here ensures each predicate
-                // evaluation starts from a known-clean state.
-                threadLocals.transitionOverrideValue = nil
-
-                failures = []
-                passedAccesses = []
-
-                context.eventsSent.removeAll()
-                context.probes.removeAll()
-
-                // Snapshot valueUpdates entry counts before predicate evaluation.
-                // Paths that gain new entries during evaluation had side-effect writes
-                // (e.g. `accessCount += 1` inside a computed getter). These must be
-                // skipped in isEqualIncludingIds because capturedValue is pre-write
-                // but lastState is post-write.
-                let prePredicateUpdateCounts: [AnyKeyPath: Int] = lock {
-                    var counts: [AnyKeyPath: Int] = [:]
-                    for (path, entries) in self.valueUpdates {
-                        counts[path] = entries.count
-                    }
-                    return counts
-                }
-
-                // Step 1: Evaluate all predicates. Reading properties inside a predicate
-                // fires willAccess → appends Access entries to context.accesses.
-                for predicate in predicates {
-                    context.predicate = predicate
-                    
-                    context.accesses.removeAll(keepingCapacity: true)
-                    context.eventsNotSent.removeAll(keepingCapacity: true)
-                    context.modelsNoLongerPartOfTester.removeAll(keepingCapacity: true)
-
-                    let passed = usingActiveAccess(self) { predicate.predicate() }
-                    let accesses = context.accesses
-                    if passed {
-                        passedAccesses += accesses
-                    } else {
-                        failures.append(.init(
-                            predicate: predicate,
-                            accesses: accesses,
-                            events: context.eventsNotSent,
-                            modelsNoLongerPartOfTester: context.modelsNoLongerPartOfTester,
-                            probes: context.probes
-                        ))
-
-                        break
-                    }
-                }
-
-                if failures.isEmpty {
-                    // Step 2: All predicates passed. Verify that the predicate's read values
-                    // match lastState (the live model). This guards against the predicate
-                    // passing transiently while a backgroundCall batch is still committing
-                    // changes (IDs would diverge because frozenCopy includes generation IDs).
-                    //
-                    // In transitions mode, accesses that read from the FIFO queue (historical
-                    // values) are skipped here — they don't correspond to live model values
-                    // and comparing them against lastState would always fail.
-                    //
-                    // frozenCopy walks Optional<M> children which calls _swift_getKeyPath to
-                    // create dynamic ContainerCursor key paths. _swift_getKeyPath needs the Swift
-                    // runtime lock — calling it while holding NSRecursiveLock deadlocks when the
-                    // runtime lock is contended. Compute frozenCopy outside the lock; expectedState
-                    // and lastState are value-type copies so this is safe.
-                    // isApplyingSnapshot suppresses willAccessPreference/willAccessStorage callbacks
-                    // and re-entrant observation triggered by releasing live child models during
-                    // the transform.
-                    // Read lastState under the lock (brief), then compute frozenCopy and
-                    // evaluate key paths OUTSIDE the lock. Evaluating container cursor key
-                    // paths (e.g. items[0].value) can call _swift_getKeyPath which acquires
-                    // the Swift runtime lock. Holding TestAccess lock while needing the
-                    // runtime lock deadlocks when another thread holds the runtime lock and
-                    // needs TestAccess lock (lock-ordering inversion). Since `last` is a
-                    // local frozen copy and `passedAccesses` is a local array, no lock is
-                    // needed for the comparison itself.
-                    let isEqualIncludingIds = threadLocals.withValue(true, at: \.isApplyingSnapshot) {
-                        // Snapshot all shared state under a single brief lock, then call
-                        // frozenCopy and run the reduce entirely OUTSIDE the lock.
-                        // frozenCopy walks Optional<M> children via _swift_getKeyPath which
-                        // acquires the Swift runtime lock — holding the TestAccess lock at the
-                        // same time deadlocks (lock-ordering inversion). The new conditions
-                        // below also read exhaustivity/valueUpdates (shared state), so we
-                        // capture those under the same brief lock.
-                        let (snap, capturedExhaustivity, capturedValueUpdates) = lock { (lastState, exhaustivity, valueUpdates) }
-                        let last = snap.frozenCopy
-                        return threadLocals.withValue(true, at: \.includeImplicitIDInMirror) {
-                            return passedAccesses.reduce(true) { result, access in
-                                // Context/preference storage values live in AnyContext.contextStorage,
-                                // not in the @Model struct fields. Writing them back to a frozen copy
-                                // (which has no live context) is a silent no-op, so the round-trip
-                                // read would return defaultValue instead of the asserted value —
-                                // making isEqualIncludingIds always false and causing a hang.
-                                // The predicate already verified these values on the live model,
-                                // so we skip the frozen-copy equality check for them.
-                                if access.skipEqualityCheck { return result }
-                                // Skip container-level model accesses. When a model property (e.g.
-                                // TestHelper.summary of type SummaryFeature) is accessed, its full
-                                // struct value is captured including the generation counter. But the
-                                // generation counter advances on every child write, so comparing the
-                                // whole struct causes a false "not settled yet" when the child property
-                                // we actually asserted has already settled. Child-level accesses
-                                // (e.g. SummaryFeature.destination) are always recorded alongside the
-                                // parent, and they check the actual leaf value with IDs — sufficient
-                                // to detect genuine backgroundCall in-flight conditions.
-                                if access.isModelTypeValue { return result }
-                                // Skip ModelContainer-typed values (Optional<M>, @ModelContainer enums).
-                                // These use ContainerCursor key paths whose getter does `get($0)!`.
-                                // ContainerCursor paths are safe to traverse only on the live model
-                                // hierarchy — reading them on a frozenCopy snapshot crashes because
-                                // frozenCopy transforms model identity and the cursor no longer matches.
-                                // Leaf accesses recorded within the container are checked independently.
-                                if access.isContainerTypeValue { return result }
-                                // Transitions mode: skip settlement check for accesses that read
-                                // from the FIFO queue. Their captured value is historical (the
-                                // front-of-queue value) and will not match lastState (the live value).
-                                if capturedExhaustivity.contains(.transitions) && capturedValueUpdates[access.path] != nil { return result }
-                                // Skip accesses whose path was WRITTEN during predicate
-                                // evaluation (e.g. `accessCount += 1` inside a computed getter).
-                                // capturedValue captures the pre-write value, but lastState
-                                // reflects the post-write value, so they'd always diverge.
-                                // Only skip paths that gained NEW entries during this predicate
-                                // cycle — paths written before the predicate still need checking.
-                                if let currentCount = capturedValueUpdates[access.path]?.count,
-                                   currentCount > (prePredicateUpdateCounts[access.path] ?? 0) { return result }
-                                // Compare the predicate-time captured value against lastState directly.
-                                // Using only `last` (which reflects the current live state) and comparing
-                                // against the captured value is safe and sufficient: if the IDs match,
-                                // the model has settled.
-                                let a = last[keyPath: access.path]
-                                return result && (diff(access.capturedValue(), a) == nil)
-                            }
-                        }
-                    }
-
-                    // Predicate passed but IDs don't match yet: there is a pending
-                    // backgroundCall batch in flight. Wait for items currently queued to
-                    // drain (FIFO sentinel), then retry. We use waitForCurrentItems rather
-                    // than waitUntilIdle so we don't block on other tests' work under 100x load.
-                    // If the timeout has elapsed even here, fall through to the failure path so
-                    // the assert doesn't spin forever when isEqualIncludingIds is permanently false.
-                    if !isEqualIncludingIds {
-                        if start.distance(to: monotonicNanoseconds()) > hardCap {
-                            // Hard cap hit: the model has been continuously producing changes
-                            // and IDs never converged. Report as failure.
-                            fail("State did not settle: model IDs kept diverging after the predicate passed. This may indicate a backgroundCallQueue loop or an unresolvable ID mismatch.", at: fileAndLine)
-                            return
-                        }
-                        await backgroundCall.waitForCurrentItems(deadline: start + hardCap)
-                        await yieldToScheduler()
-                        // Count as progress — backgroundCall draining counts as activity.
-                        lastProgressTime = monotonicNanoseconds()
-                        continue
-                    }
-
-                    if isEqualIncludingIds {
-                        // Settling mode: after the predicate passes and IDs converge,
-                        // wait for all activation tasks to enter their body, then run
-                        // one idle cycle to catch trailing mutations (e.g. forEach first
-                        // fires). Finally, reset selected exhaustivity categories so
-                        // subsequent expect {} calls only see changes made after settling.
-                        if let resetting = settleResetting {
-                            let settled = await waitUntilSettled(calibration: cal, at: fileAndLine)
-                            if !settled { return }
-
-                            // Reset selected exhaustivity categories.
-                            lock {
-                                if resetting.contains(.state) {
-                                    expectedState = lastState
-                                    valueUpdates.removeAll()
-                                }
-                                if resetting.contains(.local) {
-                                    dependencyMetadataUpdates.removeAll()
-                                }
-                                if resetting.contains(.environment) || resetting.contains(.preference) {
-                                    dependencyPreferenceUpdates.removeAll()
-                                }
-                                if resetting.contains(.events) {
-                                    events.removeAll()
-                                }
-                                if resetting.contains(.probes) {
-                                    for probe in probes {
-                                        probe.resetValues()
-                                    }
-                                }
-                            }
-                            return
-                        }
-
-                        // Step 3: Model has settled. Mark all asserted paths as handled,
-                        // advance expectedState, and run the exhaustion check.
-                        //
-                        // Capture valueUpdates inside the same lock that clears it.
-                        // checkExhaustion would otherwise re-read valueUpdates under a
-                        // separate lock acquisition, creating a race window where a concurrent
-                        // thread running an activeAccessCallback can write a new entry between
-                        // the clearing block's unlock and checkExhaustion's lock — producing
-                        // a spurious "State not exhausted" failure.
-                        var capturedValueUpdates: [PartialKeyPath<Root>: [ValueUpdate]]? = nil
-                        lock {
-                            // Sync expectedState to lastState for structural consistency.
-                            // Individual consumed.apply(&expectedState) is unsafe because deep
-                            // paths through container types (Optional<Child>, array elements)
-                            // use cursor keypaths whose setters force-unwrap — expectedState may
-                            // have a nil/stale container. Assigning lastState directly keeps all
-                            // container structures and IDs in sync. Remaining FIFO entries in
-                            // valueUpdates are caught by Layer 3 of the exhaustion check.
-                            expectedState = lastState
-
-                            let isTransitions = exhaustivity.contains(.transitions)
-                            for access in passedAccesses {
-                                if valueUpdates[access.path] != nil {
-                                    if isTransitions {
-                                        // Transitions mode: pop only the front (oldest unasserted) entry.
-                                        // Subsequent writes remain for the next assertion to consume.
-                                        valueUpdates[access.path]!.removeFirst()
-                                        if valueUpdates[access.path]!.isEmpty {
-                                            valueUpdates[access.path] = nil
-                                        }
-                                    } else {
-                                        // Non-transitions mode (last-write-wins): clear all entries.
-                                        // The assertion consumed the final value; intermediate writes
-                                        // do not need to be individually asserted.
-                                        valueUpdates[access.path] = nil
-                                    }
-                                }
-                                // For dependency context storage, run the additional cleanup (clears
-                                // dependencyMetadataUpdates for this storage key+context).
-                                access.additionalCleanup?()
-                            }
-
-                            for index in context.eventsSent.reversed() { events.remove(at: index) }
-
-                            for (probe, value) in context.probes {
-                                probe.consume(value)
-                            }
-
-                            // Capture valueUpdates after clearing so checkExhaustion uses
-                            // this consistent snapshot rather than re-reading under a new lock.
-                            capturedValueUpdates = valueUpdates
-                        }
-
-                        if enableExhaustionTest && !exhaustivity.contains(.transitions) {
-                            // Step 4: Check that no other state changed without being
-                            // asserted. Diffs expectedState against lastState.
-                            // IMPORTANT: Must be called outside the lock. checkExhaustion calls
-                            // diffMessage/customDump which triggers customMirror on @Model types,
-                            // which reads @ModelTracked properties, which calls willAccess, which
-                            // acquires context.rootPaths — also guarded by the same lock. If we
-                            // are suspended via an async withValue and resume on a different thread,
-                            // the NSRecursiveLock is not re-entrant across threads and deadlocks.
-                            checkExhaustion(at: fileAndLine, includeUpdates: true, capturedUpdates: capturedValueUpdates)
-                        }
-
-                        return
-                    }
-                }
-
-                // Step 5 (timeout): Report failures.
-                // IMPORTANT: All reporting (diffMessage, customDump, fail) happens outside the lock.
-                // diffMessage/customDump triggers customMirror on @Model types, which reads
-                // @ModelTracked properties, which calls willAccess → rootPaths → tries to acquire
-                // this same lock from the continuation's thread — deadlock if already held.
-                //
-                // Activity-relative timeout: fail only when the model has made no progress for
-                // `scaledTimeout` (no state changes), or when the absolute hard cap fires.
-                let now = monotonicNanoseconds()
-                if lastProgressTime.distance(to: now) > scaledTimeout || start.distance(to: now) > hardCap {
-                    // Build all failure messages outside the lock so that diffMessage/customDump
-                    // cannot re-enter it via willAccess → rootPaths.
-                    var reports: [[(String, FileAndLine)]] = []
-                    for failure in failures {
-                        var messages: [(String, FileAndLine)] = []
-                        if let (lhs, rhs) = failure.predicate.values() {
-                            let propertyNames = failure.accesses.compactMap { access in
-                                access.propertyName.map { "\(access.modelName).\($0)" }
-                            }.joined(separator: ", ")
-
-                            let title = "Expectation not met: \(propertyNames)"
-                            let message = threadLocals.withValue(true, at: \.includeChildrenInMirror) {
-                                diffMessage(expected: rhs, actual: lhs, title: title)
-                            }
-                            messages.append((message ?? title, failure.predicate.fileAndLine))
-                        } else {
-                            for access in failure.accesses {
-                                let pred = access.propertyName.map {
-                                    "\(access.modelName).\($0) == \(access.value())"
-                                } ?? access.value()
-                                messages.append(("Expectation not met: \(pred)", failure.predicate.fileAndLine))
-                            }
-                        }
-
-                        for event in failure.events {
-                            messages.append(("Expected event not sent: \(String(customDumping: event.event)) from \(event.context.typeDescription)", failure.predicate.fileAndLine))
-                        }
-
-                        for modelName in failure.modelsNoLongerPartOfTester {
-                            messages.append(("Model \(modelName) is no longer part of this tester", fileAndLine))
-                        }
-
-                        for (probe, value) in failure.probes {
-                            let preTitle = "Expected probe not called" + (probe.name.map { " \"\($0)\":" } ?? ":")
-                            let title = value is NoArgs ? preTitle :
-                                """
-                                \(preTitle)
-                                    \(String(customDumping: value))
-                                """
-
-                            if probe.isEmpty {
-                                messages.append((
-                                    """
-                                    \(title)
-
-                                    No available probe values
-                                    """, fileAndLine))
-                            } else if probe.count == 1, let message = threadLocals.withValue(true, at: \.includeChildrenInMirror, perform: { diffMessage(expected: value, actual: probe.values[0], title: "Probe does not match") }) {
-                                messages.append((message, fileAndLine))
-                            } else {
-                                messages.append((
-                                    """
-                                    \(title)
-
-                                    \(probe.count) Available probe values to assert:
-                                        \(probe.values.map { String(customDumping: $0) }.joined(separator: "\n    "))
-                                    """, fileAndLine))
-                            }
-                        }
-
-                        if failure.accesses.isEmpty, failure.events.isEmpty, failure.modelsNoLongerPartOfTester.isEmpty, failure.probes.isEmpty {
-                            messages.append(("Assertion failed", failure.predicate.fileAndLine))
-                        }
-                        reports.append(messages)
-                    }
-
-                    // Now apply state mutations under the lock (no diffMessage/customDump here).
-                    lock {
-                        // Sync expectedState for structural consistency (see Step 3 comment).
-                        expectedState = lastState
-
-                        for access in passedAccesses {
-                            // Pop front entry from FIFO queue if present.
-                            if valueUpdates[access.path] != nil {
-                                valueUpdates[access.path]!.removeFirst()
-                                if valueUpdates[access.path]!.isEmpty {
-                                    valueUpdates[access.path] = nil
-                                }
-                            }
-                            access.additionalCleanup?()
-                        }
-
-                        for index in context.eventsSent.reversed() { events.remove(at: index) }
-
-                        for (probe, value) in context.probes {
-                            probe.consume(value)
-                        }
-                    }
-
-                    // Emit failure reports outside the lock.
-                    for messages in reports {
-                        for (message, location) in messages {
-                            fail(message, at: location)
-                        }
-                    }
-
-
-                    if enableExhaustionTest {
-                        checkExhaustion(at: fileAndLine, includeUpdates: false)
-                    }
-                    return
-                }
-
-                // Step 6: Wait for the next modification event before re-checking.
-                // Pass the remaining no-progress budget as the wait timeout (how long to sleep
-                // if no modification arrives). The hard cap is checked at Step 5 next iteration.
-                let elapsed = lastProgressTime.distance(to: monotonicNanoseconds())
-                let remaining = elapsed < scaledTimeout ? scaledTimeout - UInt64(elapsed) : 0
-                retryCount += 1
-
-                let didProgress = await waitForModification(timeoutNanoseconds: remaining, yieldRoundNs: yieldRoundNs, retryCount: retryCount)
-                if didProgress { lastProgressTime = monotonicNanoseconds() }
-            }
-
-            // Step 5 (timeout): Report failures.
-            // IMPORTANT: All reporting (diffMessage, customDump, fail) happens outside the lock.
-            // diffMessage/customDump triggers customMirror on @Model types, which reads
-            // @ModelTracked properties, which calls willAccess → rootPaths → tries to acquire
-            // this same lock from the continuation's thread — deadlock if already held.
-            // Build all failure messages outside the lock so that diffMessage/customDump
-            // cannot re-enter it via willAccess → rootPaths.
-            var reports: [[(String, FileAndLine)]] = []
-            for failure in failures {
-                var messages: [(String, FileAndLine)] = []
-                if let (lhs, rhs) = failure.predicate.values() {
-                    let propertyNames = failure.accesses.compactMap { access in
-                        access.propertyName.map { "\(access.modelName).\($0)" }
-                    }.joined(separator: ", ")
-
-                    let title = "Expectation not met: \(propertyNames)"
-                    let message = threadLocals.withValue(true, at: \.includeChildrenInMirror) {
-                        diffMessage(expected: rhs, actual: lhs, title: title)
-                    }
-                    messages.append((message ?? title, failure.predicate.fileAndLine))
-                } else {
-                    for access in failure.accesses {
-                        let pred = access.propertyName.map {
-                            "\(access.modelName).\($0) == \(access.value())"
-                        } ?? access.value()
-                        messages.append(("Expectation not met: \(pred)", failure.predicate.fileAndLine))
-                    }
-                }
-
-                for event in failure.events {
-                    messages.append(("Expected event not sent: \(String(customDumping: event.event)) from \(event.context.typeDescription)", failure.predicate.fileAndLine))
-                }
-
-                for modelName in failure.modelsNoLongerPartOfTester {
-                    messages.append(("Model \(modelName) is no longer part of this tester", fileAndLine))
-                }
-
-                for (probe, value) in failure.probes {
-                    let preTitle = "Expected probe not called" + (probe.name.map { "\"\($0)\":" } ?? ":")
-                    let title = value is NoArgs ? preTitle :
-                        """
-                        \(preTitle)
-                            \(String(customDumping: value))
-                        """
-
-                    if probe.isEmpty {
-                        messages.append((
-                            """
-                            \(title)
-
-                            No available probe values
-                            """, fileAndLine))
-                    } else if probe.count == 1, let message = threadLocals.withValue(true, at: \.includeChildrenInMirror, perform: { diffMessage(expected: value, actual: probe.values[0], title: "Probe does not match") }) {
-                        messages.append((message, fileAndLine))
-                    } else {
-                        messages.append((
-                            """
-                            \(title)
-
-                            \(probe.count) Available probe values to assert:
-                                \(probe.values.map { String(customDumping: $0) }.joined(separator: "\n\t"))
-                            """, fileAndLine))
-                    }
-                }
-
-                if failure.accesses.isEmpty, failure.events.isEmpty, failure.modelsNoLongerPartOfTester.isEmpty, failure.probes.isEmpty {
-                    messages.append(("Expectation not met", failure.predicate.fileAndLine))
-                }
-                reports.append(messages)
-            }
-
-            // Now apply state mutations under the lock (no diffMessage/customDump here).
-            lock {
-                // Sync expectedState for structural consistency (see Step 3 comment).
-                expectedState = lastState
-
-                for access in passedAccesses {
-                    // Pop front entry from FIFO queue if present.
-                    if valueUpdates[access.path] != nil {
-                        valueUpdates[access.path]!.removeFirst()
-                        if valueUpdates[access.path]!.isEmpty {
-                            valueUpdates[access.path] = nil
-                        }
-                    }
-                    access.additionalCleanup?()
-                }
-
-                for index in context.eventsSent.reversed() { events.remove(at: index) }
-
-                for (probe, value) in context.probes {
-                    probe.consume(value)
-                }
-            }
-
-            // Emit failure reports outside the lock.
-            for messages in reports {
-                for (message, location) in messages {
-                    fail(message, at: location)
-                }
-            }
-
-            if enableExhaustionTest {
-                checkExhaustion(at: fileAndLine, includeUpdates: false)
-            }
-        }
-    }
-
-    /// Polls `expression` until it returns a non-nil value, then returns the unwrapped result.
-    /// Uses the same activity-relative idle detection as `expect`: the timeout resets on every
-    /// model state change, so a healthy model never hits it. The hard cap is the absolute
-    /// safety net (default 5 s, overridable via `TestAccessOverrides.hardCapNanoseconds`).
-    func require<T>(_ expression: @escaping @Sendable () -> T?, at fileAndLine: FileAndLine) async throws -> T {
-        let cal = await calibrate()
-        let scaledTimeout = cal.scaledTimeout
-        let yieldRoundNs = cal.yieldRoundNs
-        let hardCap = cal.hardCap
-        let start = cal.start
-        var lastProgressTime = start
-        var retryCount = 0
-
-        while true {
-            if let value = expression() {
-                // Expression is non-nil. Drive the same settled-state stability check as expect.
-                let predicate = AssertBuilder.Predicate(predicate: { expression() != nil }, fileAndLine: fileAndLine)
-                await expect(timeoutNanoseconds: nanosPerSecond, at: fileAndLine, predicates: [predicate], enableExhaustionTest: false)
-                return value
-            }
-
-            let now = monotonicNanoseconds()
-            if lastProgressTime.distance(to: now) > scaledTimeout || start.distance(to: now) > hardCap {
-                fail("Failed to unwrap value of type \(T.self)", at: fileAndLine)
-                throw UnwrapError()
-            }
-
-            let elapsed = start.distance(to: now)
-            let remaining = elapsed < hardCap ? hardCap - UInt64(elapsed) : 0
-            let didProgress = await waitForModification(
-                timeoutNanoseconds: min(remaining, yieldRoundNs),
-                yieldRoundNs: yieldRoundNs,
-                retryCount: retryCount
-            )
-            if didProgress { lastProgressTime = monotonicNanoseconds() }
-            retryCount += 1
-        }
-    }
-
-    /// Suspends briefly so the async pipeline can settle, then returns so the outer
-    /// assert loop can re-check its predicate.
-    ///
-    /// Uses an escalation strategy based on `retryCount` to balance two competing needs:
-    ///
-    /// - **Fast conditions** (e.g. a model property set by a cancellation handler):
-    ///   The value is written directly to `lastState` synchronously. A short kernel
-    ///   timer is sufficient — no need to wait for `backgroundCall` at all.
-    ///
-    /// - **Async conditions** (memoized properties, `Observed` streams):
-    ///   Both go through `backgroundCall` — memoize's performUpdate and Observed stream
-    ///   updates all use the same per-test task-local queue. A FIFO sentinel via
-    ///   `waitForCurrentItems` ensures the update has run before re-checking.
-    ///
-    /// Escalation by retry count:
-    ///   0-1: Short kernel timer (1ms) — fast path for already-settled conditions.
-    ///   2+:  Use `waitForCurrentItems(deadline:)` then `waitUntilIdle()`. Ensures
-    ///        the drain loop has run so stream consumers have had a scheduler turn.
-    ///
-    /// All waiting is done via `onAnyModification` callbacks and timers (kernel-level on
-    /// platforms with `libdispatch`, `Task.sleep` on WASM). `Task.yield()` alone is not used
-    /// for the timer because on a saturated cooperative pool it can suspend for minutes.
-    ///
-    /// Returns `true` if a state modification was observed during the wait (progress was made).
-    @discardableResult
-    package func waitForModification(timeoutNanoseconds remaining: UInt64, yieldRoundNs: UInt64, retryCount: Int = 0) async -> Bool {
-        guard remaining > 0 else { return false }
-
-        let bgQueue = backgroundCall
-
-        // A continuation slot shared between the onAnyModification callback and the
-        // DispatchQueue timer. Protected by LockIsolated so exactly one of them resumes
-        // the continuation (the first one sets slot to nil, preventing a double-resume).
-        let contSlot = LockIsolated<CheckedContinuation<Void, Never>?>(nil)
-        let didModify = LockIsolated(false)
-
-        // Resumes the continuation exactly once. Called from the onAnyModification
-        // callback (on the model-writing thread) and from the DispatchQueue timer.
-        let signal: @Sendable () -> Void = {
-            didModify.setValue(true)
-            contSlot.withValue { slot in
-                slot?.resume()
-                slot = nil
-            }
-        }
-
-        // Register BEFORE any queue drain so modifications during drain are captured.
-        let cancelModification = context.onAnyModification { _ in signal }
-        defer { cancelModification() }
-
-        // For retryCount >= 2 with pending queue items, drain the queue first.
-        // Any modifications that arrive during the drain are captured via signal().
-        if retryCount >= 2 && !bgQueue.isIdle {
-            // FIFO sentinel: suspend until all items currently in the queue have been
-            // processed. When it fires, any pending performUpdate has already run.
-            //
-            // Deadline: use the full remaining timeout. By retry 2, fast conditions
-            // (e.g. testCancelInFlight) have already passed on earlier retries, so we
-            // know this is a genuinely async condition.
-            let deadline = monotonicNanoseconds() + remaining
-            await bgQueue.waitForCurrentItems(deadline: deadline)
-            // waitUntilIdle() ensures the drain loop has fully settled so stream
-            // consumers have had a scheduling opportunity before we re-check.
-            if !bgQueue.isIdle {
-                await bgQueue.waitUntilIdle(deadline: deadline)
-            }
-            // If a modification was signalled during the drain, return now without
-            // starting another kernel timer.
-            if didModify.value { return true }
-        }
-
-        // Wait for either a modification signal or a kernel timer (whichever fires first).
-        // Timer delay:
-        // - Early retries (0-1): 1ms — fast path for already-satisfied conditions.
-        // - Later retries (2+): yieldRoundNs — avoids busy-spinning while waiting for
-        //   an async modification (e.g. forEach callback writing canUndo/canRedo).
-        if !didModify.value {
-            let delayNs: UInt64 = retryCount < 2 ? 1_000_000 : yieldRoundNs
-            await withCheckedContinuation { (cont: CheckedContinuation<Void, Never>) in
-                let alreadyModified = contSlot.withValue { slot -> Bool in
-                    if didModify.value { return true }
-                    slot = cont
-                    return false
-                }
-                if alreadyModified {
-                    cont.resume()
-                    return
-                }
-                // Timer to wake the continuation after delayNs.
-                scheduleAfter(nanoseconds: delayNs) {
-                    contSlot.withValue { slot in
-                        slot?.resume()
-                        slot = nil
-                    }
-                }
-            }
-        }
-
-        return didModify.value
-    }
-
-    // MARK: - TaskLifecycleDelegate
-
-    func activationTaskCreated() {
-        lock { _activationTasksInFlight += 1 }
-    }
-
-    func activationTaskEntered() {
-        lock { _activationTasksInFlight -= 1 }
-    }
-
-    func taskCreated() {
-        lock { _activeTaskCount += 1 }
-    }
-
-    func taskCompleted() {
-        lock { _activeTaskCount -= 1 }
-    }
-
-    /// True when all tasks born from `onActivate()` have begun executing their body.
-    var activationTasksInFlight: Int {
-        lock { _activationTasksInFlight }
-    }
-
-    /// True when no tasks are currently running anywhere in the hierarchy.
-    var isCompletelyIdle: Bool {
-        lock { _activeTaskCount == 0 }
-    }
-
-    /// Builds a diagnostic string for settle() timeout failures.
-    private func settleTimeoutDiagnostics() -> String {
-        var lines: [String] = []
-        let taskInfos = context.activeTasks
-        for info in taskInfos {
-            for (taskName, fl) in info.tasks {
-                lines.append("  \(info.modelName): \"\(taskName)\" @ \(fl.fileID):\(fl.line)")
-            }
-        }
-        return lines.joined(separator: "\n")
-    }
-
-    // MARK: - Shared calibration and settling
-
-    /// Measures scheduler latency and computes adaptive timeouts.
-    ///
-    /// The calibration yield measures how long a `yieldToScheduler()` takes under current
-    /// system load, then scales all timeouts proportionally. This makes the wait loops
-    /// robust under both light (single test) and heavy (100× parallel) conditions.
-    ///
-    /// - Parameter timeout: Base timeout for predicate waiting (default 1 s). Short explicit
-    ///   timeouts (e.g. from output snapshot tests) are preserved as-is.
     // Last scheduler-latency measurement (nanoseconds). Updated by calibrate().
     // Shared safely: calibrate() is only called from the test task (never from model
     // tasks), so writes are sequential within one TestAccess instance.
     // Default: 1ms — a reasonable assumption before the first measurement.
-    private var _lastYieldLatencyNs: UInt64 = 1_000_000
+    var _lastYieldLatencyNs: UInt64 = 1_000_000
 
-    package func calibrate(timeout: UInt64 = nanosPerSecond) async -> WaitCalibration {
-        let calibrationStart = monotonicNanoseconds()
-        await yieldToScheduler()
-        let yieldLatencyNs = monotonicNanoseconds() - calibrationStart
-        lock { _lastYieldLatencyNs = yieldLatencyNs }
-        return makeCalibration(yieldLatencyNs: yieldLatencyNs, timeout: timeout)
-    }
-
-    /// Builds a `WaitCalibration` from the most recently measured scheduler latency,
-    /// without performing a new GCD hop. Intended for end-of-test cleanup paths that
-    /// are called by hundreds of parallel tests simultaneously — a fresh GCD hop from
-    /// each would saturate the global queue and add seconds of overhead to every test.
-    package func calibrateWithCachedLatency(timeout: UInt64 = nanosPerSecond) -> WaitCalibration {
-        let yieldLatencyNs = lock { _lastYieldLatencyNs }
-        return makeCalibration(yieldLatencyNs: yieldLatencyNs, timeout: timeout)
-    }
-
-    private func makeCalibration(yieldLatencyNs: UInt64, timeout: UInt64) -> WaitCalibration {
-        let scaledTimeout = timeout >= nanosPerSecond
-            ? min(10 * nanosPerSecond, max(timeout, yieldLatencyNs * 100))
-            : timeout
-        let yieldRoundNs = max(1_000_000, min(500_000_000, yieldLatencyNs))
-        let hardCap = TestAccessOverrides.hardCapNanoseconds
-            ?? min(30 * nanosPerSecond, max(5 * nanosPerSecond, scaledTimeout * 10))
-        return WaitCalibration(
-            yieldLatencyNs: yieldLatencyNs,
-            yieldRoundNs: yieldRoundNs,
-            scaledTimeout: scaledTimeout,
-            hardCap: hardCap,
-            start: monotonicNanoseconds()
-        )
-    }
-
-    /// Waits for the model hierarchy to become idle using adaptive calibration.
-    ///
-    /// Phase 1: Wait for all `onActivate()`-born tasks to begin executing their body.
-    /// Phase 2: Idle cycle — wait until `modificationCount` stabilizes across a full
-    /// scheduling round with no active tasks running.
-    ///
-    /// Used by `expect`'s settle path and by end-of-test `checkExhaustion`.
-    /// Returns `true` on success, `false` if the hard cap was hit.
-    /// - Parameter reportTimeout: When `true` (default), reports a test failure if the
-    ///   hard cap is hit. Pass `false` from `checkExhaustion` where the subsequent
-    ///   exhaustivity check handles failure reporting through proper exhaustivity bits.
-    @discardableResult
-    package func waitUntilSettled(calibration cal: WaitCalibration, reportTimeout: Bool = true, at fileAndLine: FileAndLine) async -> Bool {
-        // Phase 1: Wait for all activation tasks to enter their body.
-        // In practice this completes almost immediately because activationTaskEntered()
-        // fires at the start of the task body.
-        //
-        // Skip when called from checkExhaustion (reportTimeout: false): activation tasks
-        // that haven't started yet have already been cancelled by cancelAllRecursively()
-        // and removed from Cancellations.registered / activeTasks. Waiting for them to
-        // start under 500+ parallel test load can take 30 s (cooperative pool saturation
-        // causes 30,000 × 1ms iterations). The subsequent checkExhaustion(checkTasks:)
-        // call won't see them — safe to skip.
-        if reportTimeout {
-            while activationTasksInFlight > 0 {
-                await waitForModification(timeoutNanoseconds: cal.scaledTimeout, yieldRoundNs: cal.yieldRoundNs, retryCount: 0)
-                if cal.start.distance(to: monotonicNanoseconds()) > cal.hardCap {
-                    let taskInfo = settleTimeoutDiagnostics()
-                    fail("settle() timed out: model still has active tasks.\n\(taskInfo)", at: fileAndLine)
-                    return false
+    /// Resets exhaustivity categories within the lock. Called from settle(resetting:) settle paths in expect().
+    func _applyResetting(_ bits: _ExhaustivityBits) {
+        lock {
+            if bits.contains(.state) {
+                expectedState = lastState
+                valueUpdates.removeAll()
+            }
+            if bits.contains(.local) {
+                dependencyMetadataUpdates.removeAll()
+            }
+            if bits.contains(.environment) || bits.contains(.preference) {
+                dependencyPreferenceUpdates.removeAll()
+            }
+            if bits.contains(.events) {
+                events.removeAll()
+            }
+            if bits.contains(.probes) {
+                for probe in probes {
+                    probe.resetValues()
                 }
             }
         }
-
-        // Phase 2: Idle cycle — wait until one full scheduling round passes
-        // with no state changes. Uses modificationCount on the root context as
-        // a version number.
-        var lastChangeVersion = context.modificationCount
-        // Tracks whether we already confirmed _activeTaskCount == 0 with a stable
-        // modificationCount in the previous outer-loop iteration.
-        //
-        // The race we guard against: a just-completed task dispatches its final model
-        // mutation to backgroundCall AND decrements _activeTaskCount in the same
-        // cooperative-pool turn. The next check sees _activeTaskCount == 0 and a stable
-        // modificationCount (backgroundCall hasn't committed the mutation yet), and
-        // would break prematurely — baseline captures the pre-mutation value.
-        //
-        // Fix: require two consecutive outer-loop iterations that both see
-        // _activeTaskCount == 0 with no modificationCount change. The
-        // waitForCurrentItems at the START of the NEXT iteration (line 1332) naturally
-        // drains the pending mutation, so modificationCount changes and we loop once
-        // more to re-confirm. No extra drain is added; the existing loop structure
-        // handles it. For observation-heavy models this adds at most two cheap
-        // no-op iterations rather than triggering cascading observation re-fires.
-        var sawZeroActiveTasks = false
-        while true {
-            await backgroundCall.waitForCurrentItems(deadline: cal.start + cal.hardCap)
-            await yieldToScheduler()
-
-            // Re-check activation tasks: backgroundCall drain or yieldToScheduler
-            // may have triggered child model activations (e.g. SearchResultItem
-            // activated when results is set), creating new tasks that haven't entered
-            // their body yet. Wait for them before evaluating idle state.
-            // Skipped when called from checkExhaustion (reportTimeout: false) for the
-            // same reason as Phase 1: cancelled tasks that haven't started yet are safe
-            // to ignore (removed from activeTasks by cancelAllRecursively).
-            if reportTimeout {
-                while activationTasksInFlight > 0 {
-                    await waitForModification(timeoutNanoseconds: cal.scaledTimeout, yieldRoundNs: cal.yieldRoundNs, retryCount: 0)
-                    if cal.start.distance(to: monotonicNanoseconds()) > cal.hardCap {
-                        let taskInfo = settleTimeoutDiagnostics()
-                        fail("settle() timed out: model still has active tasks.\n\(taskInfo)", at: fileAndLine)
-                        return false
-                    }
-                }
-            }
-
-            let currentVersion = context.modificationCount
-            if currentVersion == lastChangeVersion {
-                if lock({ _activeTaskCount }) == 0 {
-                    if sawZeroActiveTasks {
-                        break // confirmed: two consecutive checks with no active tasks and no changes
-                    }
-                    sawZeroActiveTasks = true
-                    continue // loop once more; waitForCurrentItems at top drains any pending mutations
-                }
-                sawZeroActiveTasks = false
-                // Tasks are still running but no state changed yet. Use a GCD-backed
-                // timer (waitForModification) rather than Task.yield() loops. Under
-                // 500+ parallel tests the cooperative pool is saturated — each
-                // Task.yield() can take 1-2s, so 20 yields × 1.5s = 30s = hardCap.
-                // waitForModification uses a DispatchQueue timer and an
-                // onAnyModification callback, so it wakes immediately on any write
-                // regardless of cooperative pool pressure.
-                //
-                // Patience is context-dependent:
-                //
-                // reportTimeout: true  (active test — tasks are still running and may write)
-                //   max(yieldRoundNs × 3, 300ms): scales with load. Under saturation
-                //   (yieldRoundNs ~500ms) a completing task may need 1–2 cooperative pool
-                //   turns (~500ms–1s) before it can write; 1.5s ensures we don't break
-                //   out before the write arrives.
-                //
-                // reportTimeout: false (checkExhaustion cleanup — tasks are cancelled)
-                //   min(yieldRoundNs × 30, 300ms): always ≤ 300ms. Cancelled tasks won't
-                //   write again; we only need to detect observation loops. The outer
-                //   yieldToScheduler() loop then gives cancelled tasks cooperative-pool
-                //   turns to exit and decrement _activeTaskCount without the 1.5s
-                //   per-test overhead that accumulates across 500+ parallel tests.
-                let patience = reportTimeout
-                    ? max(cal.yieldRoundNs * 3, 300_000_000)
-                    : min(cal.yieldRoundNs * 30, 300_000_000)
-                await waitForModification(timeoutNanoseconds: patience, yieldRoundNs: patience, retryCount: 2)
-                if context.modificationCount == currentVersion {
-                    break // No progress — remaining tasks are observation loops or cancelled tasks
-                }
-                lastChangeVersion = context.modificationCount
-                sawZeroActiveTasks = false
-            } else {
-                lastChangeVersion = currentVersion
-                sawZeroActiveTasks = false
-            }
-            if cal.start.distance(to: monotonicNanoseconds()) > cal.hardCap {
-                if reportTimeout {
-                    let taskInfo = settleTimeoutDiagnostics()
-                    fail("settle() timed out: model still has active tasks.\n\(taskInfo)", at: fileAndLine)
-                }
-                return false
-            }
-        }
-
-        return true
     }
 
-    // Checks that no state changed without being asserted (exhaustion check).
-    //
-    // At the end, resets expectedState = lastState so the next assert starts from a
-    // clean baseline.
     func checkExhaustion(at fileAndLine: FileAndLine, includeUpdates: Bool, checkTasks: Bool = false, capturedUpdates: [PartialKeyPath<Root>: [ValueUpdate]]? = nil) {
         if checkTasks {
             for info in context.activeTasks {
